Move macro logic into util fns
